requirements.txt
django
gunicorn
django-heroku
psycopg2-binary
django-cockroachdb